refactor: add product dto and payment dto
