+++ src/order/dtos/payment.dto.ts
@@ -1,0 +1,8 @@
+import { PaymentTypeEnum } from '../types/payments-types-enum';
+
+export class PaymentDto {
+  cardholder: string;
+  paymentStatus: PaymentTypeEnum;
+  value: number;
+  dueDate: Date;
+}
