+++ src/order/dtos/product.dto.ts
@@ -1,0 +1,5 @@
+export class ProductDto {
+  id: string;
+  price: number;
+  name: string;
+}
