+++ src/order/order-service.ts
@@ -1,5 +1,5 @@
 import { Injectable, Inject } from '@nestjs/common';
-import { CreateOrderDto } from './dto/order.dto';
+import { CreateOrderDto } from './dtos/order.dto';
 import { Order } from './entities/order';
 import { OrderInterface } from './interfaces/order-service.interface';
 import { OrderRepositoryInterface } from './interfaces/order-repository-interface';
@@ -12,7 +12,7 @@ export class OrderService implements OrderInterface {
     private readonly orderRepository: OrderRepositoryInterface,
   ) {}
 
-  async create(createOrderDto: CreateOrderDto) {
+  async create(createOrderDto: CreateOrderDto): Promise<Order> {
     return this.orderRepository.create(createOrderDto);
   }
