+++ src/order/order.controller.ts
@@ -8,7 +8,7 @@ import {
   Delete,
   Inject,
 } from '@nestjs/common';
-import { CreateOrderDto } from './dto/order.dto';
+import { CreateOrderDto } from './dtos/order.dto';
 import { Order } from './entities/order';
 import { OrderInterface } from './interfaces/order-service.interface';
